countAllUtentiAdmin + test

+++ manytomanyjpamaven/src/main/java/it/manytomanyjpamaven/dao/UtenteDAO.java
@@ -13,7 +13,7 @@ public interface UtenteDAO extends IBaseDAO<Utente> {
 
 	public List<Utente> findAllCreatiAGiugno2022();
 
-	public int countAllUtentiAdmin();
+	public long countAllUtentiAdmin();
 
 	public List<Utente> findAllPasswordMenoDi8Caratteri();
